renamed functiontext -> function
