+++ src/components/common/CodeSnippet/Function/index.js
