+++ src/components/common/CodeSnippet/Function/styles.js
