+++ src/components/common/CodeSnippet/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { withStyles } from "@material-ui/styles";
 
-import FunctionText from "./FunctionText";
+import Function from "./Function";
 import Key from "./Key";
 import ValueString from "./ValueString";
 import ValueNumber from "./ValueNumber";
@@ -11,7 +11,7 @@ import { useStyles } from "./styles";
 const CodeSnippet = ({ classes }) => {
   return (
     <div className={classes.codeSnippetContainer}>
-      <FunctionText text={"FunctionText"} />
+      <Function text={"FunctionText"} />
       <Key text={"Ket Text"} />
       <ValueString text={"Value String"} />
       <ValueNumber number={"Value Number"} />
